==> NoteBooks/1) TicketMaster_data.ipynb <==
##### This notebook call the Ticket Master API and save the data in a json file

# Ticket Master API- Discovery API

#import required libraries
import requests # this will be used to call the APIs
import json     # to save data to a json_file
import pandas as pd 
import os       # use this to access enviornment variable
from get_data_API import get_events


TicketMaster_key= os.getenv("TicketMasterAPI")
TM_key=os.getenv("TM_yahoo")


## Event Search for Canada
#### use the URL to get events based on the query parameter of countryCode-Canada

# define variables for the required data
country_code="CA" # code for Canada to get events
endpoint= '/discovery/v2/events' #endpoint to get event_ID
segment =["Music"] # use different segments (music,arts,sports,theatre)



res= get_events(endpoint,TM_key,country_code,segment,100) # initial page number is 0
res.status_code

# to check the number of pages of data
if res.status_code==200:
        events_data= res.json()
        if 'page' in events_data and 'totalPages' in events_data['page']:
            total_pages = events_data['page']['totalPages']

total_pages

events=[] # empty list to add each json file as an object
# to keep track of iterations use page_number
id=1000 # to label the json files 

# to give correct outputs and save the data as json file 
# call the function to get data

if res.status_code==200:
        events_data= res.json()
        if 'page' in events_data and 'totalPages' in events_data['page']:
            total_pages = events_data['page']['totalPages']
    
            for page_number in range(1000,total_pages):
                
                response = get_events(endpoint,TM_key, country_code, segment, page_number)
                data = response.json()

                if '_embedded' in data and 'events' in data["_embedded"]:
                    events.extend(data['_embedded']['events']) # adding data to existing list (empty or old data)
                 # Append events to the JSON file
                    #change file name for different parameter setting for segment or country etc. 
                    
                    with open(f'Events_{id}.json', 'w') as f: 
                                json.dump(events, f)
                    id+=1
                    print("Pages processed = ", page_number)
                else:
                    print(f"Error: No events found on page {page_number}")

           

        print("END of list") # to know end of iteration
        
else:
    print(f"Error: {res.status_code}, {res.text}")
    

  
    

# final dump to get it all in one file
#change variable name for change of segment
with open(f'Events_music.json', 'w') as f:
     json.dump(events, f)
