==> NoteBooks/4) Final_TicketMaster.ipynb <==
##### This notebook combines the data pulled from the Ticket master API into one dataframe and save the data as Final_TicketMaster.csv

import pandas as pd

#import the csv's
music_df=pd.read_csv('TicketMaster_music.csv')
sports_df=pd.read_csv('TicketMaster_sports.csv')


music_df.drop_duplicates(inplace=True)

# Combine the 2 csv's vertically stacking them 

TicketMaster_df=pd.concat([music_df,sports_df],axis=0) 

TicketMaster_df.shape

# save this to a csv
TicketMaster_df.to_csv("TicketMaster_merged.csv",index=False)

TicketMaster_df.info()

# Get unique combinations of latitude and 
latitude_venue=TicketMaster_df[['Latitude','City']].drop_duplicates()
latitude_venue.dropna(inplace=True)
#Extract unique latitude
latitude_array=latitude_venue['Latitude'].tolist()

# Get unique combinations of longitude and venue
longitude_venue=TicketMaster_df[['Longitude','City']].drop_duplicates()
longitude_venue.dropna(inplace=True)

#Extract unique longitude
longitude_array=longitude_venue['Longitude'].tolist()

venue_array=TicketMaster_df['Venues'].unique()
address_array=TicketMaster_df['Address'].unique()

# Get unique combinations of city and venue
city_venue = TicketMaster_df[['City', 'Venues']].drop_duplicates()

# Extract unique cities
city_array = city_venue['City'].tolist()

# Get unique combinations of state and venue
state_venue = TicketMaster_df[['States', 'Venues']].drop_duplicates()

# Extract unique cities
states_array = state_venue['States'].tolist()

print(len(latitude_array),len(longitude_array),len(venue_array),len(address_array),len(city_array), len(states_array))

Lat_Lon_df=pd.DataFrame({'Venue':venue_array,
                         'Address':address_array,
                         'City':city_array,
                         'State':states_array,
                         'latitude': latitude_array,
                         'longitude':longitude_array
                        })
Lat_Lon_df

Lat_Lon_df.to_csv('Venue_locations.csv',index=False)
