==> NoteBooks/5) Yelp_get_data.ipynb <==
#####  This notebook is used to call yelp API at different endpoints to get information for reviews and ratings
##### We load the Venue_locations.csv that contain the venues got from the ticketMaster API 
##### We use the latitude and longitude columns to search for business in the area
##### We use the following columns to get business match from yelp:
        Venue
        Address
        City
        State
##### from the Business_info.csv we get id that is used to get the reviews and ratings


import pandas as pd
import requests
import json
import os
Yelp_key=os.getenv('Yelp_API')
Yelp_key_MR=os.getenv('Yelp_API(MR)')
from get_data_API import get_venues_yelp
from get_data_API import get_events_yelp
from get_data_API import get_business_match
from get_data_API import get_reviews


## to verify the funtion is imported correctly
get_venues_yelp
# get_events_yelp
# get_business_match

Venue_df=pd.read_csv('Venue_locations.csv')
Venue_df.head()

unique_states = Venue_df['State'].unique()
print(unique_states)

# map the states and make the changes in place in the data frame

state_mapping = {
    'Ontario': 'ON',
    'Alberta':'AB',
    'British Columbia':'BC',
    'Manitoba':'MB',
    'Saskatchewan':'SK',
    'New Brunswick':'NB',
    'Newfoundland and Labrador':'NL',
    'Prince Edward Island':'PEI',
    'Nova Scotia':'NS',
    'Quebec':'QC'
}

Venue_df['State'] = Venue_df['State'].map(state_mapping).fillna(Venue_df['State'])
Venue_df


# to save the changes made back to the csv file 
Venue_df.to_csv('Venue_locations.csv',index=False)

## Call get_venues_yelp to get businesses in the area using latitude and longitude

# yelp_data=[]


# for index, row in Venue_df.iterrows():
#     lat=row['latitude']
#     lon=row['longitude']
#     res_yelp = get_venues_yelp(lat,lon,Yelp_key) 
#     if res_yelp.status_code == 200:
#         data = res_yelp.json()
#         if 'businesses' in data:
#             yelp_data.extend(data['businesses'])
#         # dump data into json file
#             with open(f'yelp.json', 'w') as f: 
#                     json.dump(yelp_data, f) 
#             print("Data dumped for index:", index)
#         else:
#             print("No businesses found in data for index:", index)

        

#     else:
#         print(f"Error: {res.status_code}, {res.text}")

# print("END of list") # to know end of iteration

## Call get_business_match to get business id for unique venues in the data set

# business_id=[]
# business_name=[]
# cities=[]
# for index, row in Venue_df.iterrows():
#     name=row['Venue']
#     address=row['Address']
#     city=row['City']
#     state=row['State']
#     res_id = get_business_match(name,address,city,state,Yelp_key_MR)
    
#     if res_id.status_code == 200:
#         data = res_id.json()
#         with open( f'business.json','w') as f:
#             json.dump(data,f)
#        # Check if 'businesses' key exists in the JSON response and if it's not empty
#         if 'businesses' in data and len(data['businesses']) > 0:
#             business_id.append(data['businesses'][0]['id']) # only grab business id from the json
#             business_name.append(data['businesses'][0]['name']) # only grab name from json
#             cities.append(city)
#         else:
#             business_id.append('Not available')
#             business_name.append(name)
#             cities.append(city)

#         print("Index = " ,index)
            
#     else:
#          print(f"Error: {res.status_code}, {res.text}")

# print("Done")

 print(len(data['businesses']))

# Business_df=pd.DataFrame({'business_id':business_id,
#                         'business_name':business_name,
#                          'city':cities})
# Business_df.to_csv('Business_info.csv',index=False)

Business_df.shape

## Call  get_reviews using the business id provided above in Business_df to get review and rating information 

# Business_df=pd.read_csv('Business_info.csv')
# Business_df

filtered_df = Business_df[Business_df['business_id'] != 'Not available']
filtered_df.shape

# ratings=[]
# for index, row in filtered_df.iterrows():
#     id=row['business_id']
#     res_review=get_reviews(id,Yelp_key)
#     if res_review.status_code == 200:
#         review = res_review.json()
#        # Check if 'reviews' key exists in the JSON response and if it's not empty
#         if 'reviews' in review :
#            ratings.append(review) # to keep list of reviews for each venue
#             #dump data into json file
#            with open(f'ratings_yelp1.json', 'w') as f: 
#                  json.dump(ratings, f) 
#         else:
#             business_id.append('Not available')
#             business_name.append(name)
#     else:
#          print(f"Error: {res_review.status_code}, {res_review.text}")
#          ratings.append('Not available') # for venues with no business id available

# print(res_review.status_code)
