==> NoteBooks/6) Yelp_Data 1.ipynb <==
##### Here the yelp json data is loaded that is places of interest for the venues from ticket master API and parsed and saved into  Yelp_data_1.csv

##### The Yelp_get_data.ipynb calls the API to get the json information required.
##### Once we have the business id for the restaurants we use that information to call the get_reviews function to gwt user and reviews information for each restaurant 

import pandas as pd
import json
from get_data_API import get_reviews
import os
Yelp_key=os.getenv('Yelp_API(MR)')

with open('yelp.json','r') as f:
    yelp_data_1=json.load(f)

"""
data needed from the json
id
name
rating
price
categories-[0]title
distance
location-display_Address,city,state
coordinates-latitude,longitude

"""

# create empty lists 
ID=[]
Name=[]
Ratings=[]
Pricing=[]
Category=[]
Distances=[]
Address=[]
City=[]
State=[]
Latitude=[]
Longitude=[]

for data in yelp_data_1:
    id=data['id']
    name=data['name']
    rating=data['rating']
    prices=data.get('price', 'Information not available')
    distance=data['distance']
    category=data['categories'][0]['title']
    location=data['location']['address1']
    city=data['location']['city']
    state=data['location']['state']
    latitude=data['coordinates']['latitude']
    longitude=data['coordinates']['longitude']

    ID.append(id)
    Name.append(name)
    Ratings.append(rating)
    Pricing.append(prices)
    Category.append(category)
    Distances.append(distance)
    Address.append(location)
    City.append(city)
    State.append(state)
    Latitude.append(latitude)
    Longitude.append(longitude)
    

data_dict={'Business ID': ID,
          'Name': Name,
          'Ratings': Ratings,
          'Pricing': Pricing,
          'Category':Category,
          'Distance_From_Venue':Distances,
          'Address': Address,
          'City':City,
          'State':State,
          'Latitude':Latitude,
          'Longitude': Longitude
          }

# Check lengths of all lists or arrays in the data_dict
lengths = {key: len(value) for key, value in data_dict.items()}
print(lengths)

# Create Data frame

yelp_df_1=pd.DataFrame(data=data_dict)


yelp_df_1.shape
              

yelp_df_1.drop_duplicates() # need to change ville de quebec to quebec

# store data frame to csv
yelp_df_1.to_csv('Yelp_data_1.csv',index=False)

yelp_df_1=pd.read_csv('Yelp_data_1.csv')

yelp_df_1['City'].nunique()#.tolist()

#### Use the business Id from above to get more data for reviews and ratings from the get_reviews function

ratings=[]
city=[]
category=[]
#start=464
# Specify the file path where you want to save the JSON file
file_path = 'yelp_ratings_rest.json'

for index, row in yelp_df_1.iterrows():
    id=row['Business ID']
    
    res_review=get_reviews(id,Yelp_key)
    res_review.status_code
    if index%50==0:
        print(index)
    if res_review.status_code == 200:
            review = res_review.json()
            # Check if 'reviews' key exists in the JSON response and if it's not empty
            if 'reviews' in review :
                ratings.append(review) # to keep list of reviews for each venue
                # Open the file in write mode and dump the data into it
                with open(file_path, 'w') as f:
                    json.dump(ratings, f)
                city.append(row['City'])
                category.append(row['Category'])
            else:
                ratings.append('N/A')
                city.append(row['City'])
                category.append(row['Category'])
            
    else:
             print("Stopped at= ",index)
             print(f"Error: {res_review.status_code}, {res_review.text}")
             ratings.append('N/A') # for venues with no business id available
             city.append(row['City'])
             category.append(row['Category'])
                           
print("Stopped at= ",index)


res_review.json()

data=res_review.json()
data['reviews'][0]

with open ('yelp_ratings_rest.json','r') as f:
    data=json.load(f)

data[0]
