==> NoteBooks/7) Yelp_Venue_Ratings.ipynb <==
##### In this notebook we load the json file from yelp API with the reviews and ratings information and the Business_info.csv containing id and name 
##### The json is parsed to create a new df with all the information :
            venue
            venue_id
            reviews
            ratings      
 

import pandas as pd
import json

with open('ratings_yelp1.json','r') as f:
    ratings_yelp1=json.load(f)



Business_df=pd.read_csv('Business_info.csv')
Business_df

"""
columns required from the json
reviews-text, rating
"""

ratings_yelp1[1]['reviews']#[0]['id'] #see the format of json

review=[]
ratings=[]
venue_name=[]
venue_id=[]
city=[]
for index, row in Business_df.iterrows():
    # Get the business ID from the current row
    if row['business_id'] != 'Not available':
    
    # Search for the corresponding rating in the ratings JSON data
        for review_item in ratings_yelp1:
    
            if 'reviews' in review_item :
                
                 for item in review_item['reviews']:
                     

                    if 'text' in item:
                        review.append(item['text'])
                    else:
                        review.append('N/A')

                    if 'rating' in item:
                        ratings.append(item['rating'])
                    else:
                        ratings.append('N/A')
                
                    venue_name.append(row.iloc[1])
                    venue_id.append(row.iloc[0])
                    city.append(row.iloc[2])
            # else:
            #     review.append('Not available')
            #     ratings.append('Not available')
            #     venue_name.append(row.iloc[1])
            #     venue_id.append(row.iloc[0])
            #     city.append(row.iloc[2])
    else:
            review.append('N/A')
            ratings.append('N/A')
            venue_name.append(row.iloc[1])
            venue_id.append(row.iloc[0])
            city.append(row.iloc[2])

df=pd.DataFrame({'venue':venue_name,
                 'venue_id':venue_id,
                 'city':city,
                 'reviews':review,
                 'ratings':ratings})
df
    

df=df.drop_duplicates()
df.shape

df.to_csv('Venue_Ratings.csv',index=False)
