==> NoteBooks/11 )Yelp_POI.ipynb <==
#### Here we are going to parse through the json containing information for restaurant ratings and reviews and create a new data frame to give us all the required information to built the recommender system

import pandas as pd
import json

yelp_POI=pd.read_csv('Yelp_data_1.csv')
yelp_POI.head(50)

with open('yelp_ratings_rest.json','r') as f:
    yelp_POI_ratings=json.load(f)

with open('yelp_ratings_rest1.json','r') as f1:
    yelp_POI_ratings1=json.load(f1)

with open('yelp_ratings_rest4.json','r') as f4:
    yelp_POI_ratings4=json.load(f4)

with open('yelp_ratings_rest8.json','r') as f8:
    yelp_POI_ratings8=json.load(f8)

with open('yelp_ratings_rest9.json','r') as f9:
    yelp_POI_ratings9=json.load(f9)

"""
columns required from the json files
loop through csv adn for each element in csv find yelp_POI_ratings[0]:
check for reviews - reviews , loop through reviews to get the following
rating-reviews[0]['rating']
text- reviews[0]['text']
user_id-reviews[0]['user']['id']
user_name-reviews[0]['user']['name']

"""

ratings=[]
review=[]
user_id=[]
user_name=[]
venue=[]
venue_id=[]
city=[]
categories=[]
index1=[]
index2=[]

# index starts=0,464,963,2463,#4464


row_number=0
i=4464
while i < len(yelp_POI): # start the index based on json file how data is pulled
    row=yelp_POI.iloc[i]
    index=row.name
        
    business_id=row['Business ID']
    venue_name=row['Name']
    cities=row['City']
    category=row['Category']
        
    if row_number <= len(yelp_POI_ratings9): # change json file name
        # if x==row_number:
            if 'reviews' in yelp_POI_ratings9[row_number] : # change json file name
                            
                     for item in yelp_POI_ratings9[row_number]['reviews']: # change json file name         

                        if 'text' in item:
                            review.append(item['text'])
                        else:
                            review.append('N/A')

                        if 'rating' in item:
                            ratings.append(item['rating'])
                        else:
                            ratings.append('N/A')

                        if 'user' in item:
                            user_id.append(item['user']['id'])
                            user_name.append(item['user']['name'])
                        else:
                            user_id.append('N/A')
                            user_name.append('N/A')
                
                        venue.append(venue_name)
                        venue_id.append(business_id)
                        city.append(cities)
                        index1.append(row_number)
                        index2.append(index)
                     i+=1
            
            else:
                    review.append('N/A')
                    ratings.append('N/A')
                    user_id.append('N/A')
                    user_name.append('N/A')
                    venue.append(venue_name)
                    venue_id.append(business_id)
                    city.append(cities)
                    index1.append(row_number)
                    index2.append(index)
                    
                    
            #  Stop iterating over yelp_POI_ratings once the current row is processed
            print(index) # to keep track for the different json files
    else:
        break
       
    row_number+=1
    
        
            
print(len(review),len(ratings),len(city),len(index1))

df=pd.DataFrame({'venue':venue,
                 'venue_id':venue_id,
                  'city':city,
                  'reviews':review,
                  'user_id':user_id,
                  'user_name':user_name,
                  'ratings':ratings,
                   'index':index1,
                 'index_df':index2
                })

df.shape

df.tail(50)

df.to_csv('Yelp_rest_review.csv',index=False)
